refactor(runtime): split rpc protocol runtime domains (#1685)
