+++ crates/tau-runtime/src/rpc_protocol_runtime.rs
@@ -6,17 +6,30 @@
 
 use std::{
     collections::{BTreeMap, BTreeSet, VecDeque},
-    io::{BufRead, Write},
-    path::Path,
     str::FromStr,
-    time::{SystemTime, UNIX_EPOCH},
 };
 
-use anyhow::{anyhow, bail, Context, Result};
-use serde::{Deserialize, Serialize};
+use anyhow::{anyhow, bail, Result};
+use serde::Serialize;
 use serde_json::{json, Value};
 
 use crate::rpc_capabilities_runtime::{rpc_capabilities_payload, RPC_PROTOCOL_VERSION};
+
+mod dispatch;
+mod parsing;
+mod transport;
+
+pub use dispatch::dispatch_rpc_frame_impl as dispatch_rpc_frame;
+use dispatch::{
+    dispatch_rpc_frame_for_serve_impl as dispatch_rpc_frame_for_serve,
+    lifecycle_control_audit_record_for_dispatch_impl as lifecycle_control_audit_record_for_dispatch,
+};
+pub use parsing::parse_rpc_frame_impl as parse_rpc_frame;
+pub use parsing::validate_rpc_frame_file_impl as validate_rpc_frame_file;
+pub use transport::dispatch_rpc_ndjson_input_impl as dispatch_rpc_ndjson_input;
+pub use transport::dispatch_rpc_raw_with_error_envelope_impl as dispatch_rpc_raw_with_error_envelope;
+pub use transport::serve_rpc_ndjson_reader_impl as serve_rpc_ndjson_reader;
+pub use transport::serve_rpc_ndjson_reader_with_lifecycle_audit_impl as serve_rpc_ndjson_reader_with_lifecycle_audit;
 
 pub const RPC_FRAME_SCHEMA_VERSION: u32 = 1;
 const RPC_COMPATIBLE_REQUEST_SCHEMA_VERSIONS: [u32; 2] = [0, RPC_FRAME_SCHEMA_VERSION];
@@ -238,614 +251,10 @@ fn remember_closed_run_state(
     }
 }
 
-#[derive(Debug, Clone, Deserialize)]
-struct RawRpcFrame {
-    schema_version: u32,
-    request_id: String,
-    kind: String,
-    payload: Value,
-}
-
-pub fn parse_rpc_frame(raw: &str) -> Result<RpcFrame> {
-    let frame =
-        serde_json::from_str::<RawRpcFrame>(raw).context("failed to parse rpc frame JSON")?;
-    if !RPC_COMPATIBLE_REQUEST_SCHEMA_VERSIONS.contains(&frame.schema_version) {
-        let supported_versions = RPC_COMPATIBLE_REQUEST_SCHEMA_VERSIONS
-            .iter()
-            .map(|version| version.to_string())
-            .collect::<Vec<_>>()
-            .join(", ");
-        bail!(
-            "unsupported rpc frame schema: supported request schema versions are [{}], found {}",
-            supported_versions,
-            frame.schema_version
-        );
-    }
-    let request_id = frame.request_id.trim();
-    if request_id.is_empty() {
-        bail!("rpc frame request_id must be non-empty");
-    }
-    let kind = RpcFrameKind::from_str(frame.kind.trim())?;
-    let payload = frame
-        .payload
-        .as_object()
-        .ok_or_else(|| anyhow!("rpc frame payload must be a JSON object"))?
-        .clone();
-    Ok(RpcFrame {
-        request_id: request_id.to_string(),
-        kind,
-        payload,
-    })
-}
-
-pub fn validate_rpc_frame_file(path: &Path) -> Result<RpcFrame> {
-    let raw = std::fs::read_to_string(path)
-        .with_context(|| format!("failed to read rpc frame file {}", path.display()))?;
-    parse_rpc_frame(&raw)
-}
-
-pub fn dispatch_rpc_frame(frame: &RpcFrame) -> Result<RpcResponseFrame> {
-    match frame.kind {
-        RpcFrameKind::CapabilitiesRequest => {
-            let negotiated_request_schema_version =
-                resolve_capabilities_requested_schema_version(&frame.payload)?;
-            let capabilities = rpc_capabilities_payload();
-            let capability_list = capabilities["capabilities"]
-                .as_array()
-                .cloned()
-                .ok_or_else(|| anyhow!("rpc capabilities payload is missing capabilities array"))?;
-            let contracts = capabilities["contracts"]
-                .as_object()
-                .cloned()
-                .ok_or_else(|| anyhow!("rpc capabilities payload is missing contracts object"))?;
-            Ok(build_response_frame(
-                &frame.request_id,
-                "capabilities.response",
-                json!({
-                    "protocol_version": RPC_PROTOCOL_VERSION,
-                    "response_schema_version": RPC_FRAME_SCHEMA_VERSION,
-                    "supported_request_schema_versions": RPC_COMPATIBLE_REQUEST_SCHEMA_VERSIONS,
-                    "negotiated_request_schema_version": negotiated_request_schema_version,
-                    "capabilities": capability_list,
-                    "contracts": contracts,
-                }),
-            ))
-        }
-        RpcFrameKind::RunStart => {
-            let prompt =
-                require_non_empty_payload_string(&frame.payload, "prompt", frame.kind.as_str())?;
-            let run_id = resolve_run_start_run_id(&frame.payload, &frame.request_id)?;
-            Ok(build_response_frame(
-                &frame.request_id,
-                "run.accepted",
-                json!({
-                    "status": "accepted",
-                    "mode": RPC_STUB_MODE,
-                    "prompt_chars": prompt.chars().count(),
-                    "run_id": run_id,
-                }),
-            ))
-        }
-        RpcFrameKind::RunCancel => {
-            let run_id =
-                require_non_empty_payload_string(&frame.payload, "run_id", frame.kind.as_str())?;
-            Ok(build_response_frame(
-                &frame.request_id,
-                "run.cancelled",
-                json!({
-                    "status": "cancelled",
-                    "terminal": true,
-                    "terminal_state": "cancelled",
-                    "mode": RPC_STUB_MODE,
-                    "run_id": run_id,
-                }),
-            ))
-        }
-        RpcFrameKind::RunComplete => {
-            let run_id =
-                require_non_empty_payload_string(&frame.payload, "run_id", frame.kind.as_str())?;
-            Ok(build_response_frame(
-                &frame.request_id,
-                "run.completed",
-                json!({
-                    "status": "completed",
-                    "terminal": true,
-                    "terminal_state": "completed",
-                    "mode": RPC_STUB_MODE,
-                    "run_id": run_id,
-                }),
-            ))
-        }
-        RpcFrameKind::RunFail => {
-            let run_id =
-                require_non_empty_payload_string(&frame.payload, "run_id", frame.kind.as_str())?;
-            let reason = resolve_run_fail_reason(&frame.payload)?;
-            Ok(build_response_frame(
-                &frame.request_id,
-                "run.failed",
-                json!({
-                    "status": "failed",
-                    "terminal": true,
-                    "terminal_state": "failed",
-                    "mode": RPC_STUB_MODE,
-                    "run_id": run_id,
-                    "reason": reason,
-                }),
-            ))
-        }
-        RpcFrameKind::RunTimeout => {
-            let run_id =
-                require_non_empty_payload_string(&frame.payload, "run_id", frame.kind.as_str())?;
-            let reason = resolve_run_timeout_reason(&frame.payload)?;
-            Ok(build_response_frame(
-                &frame.request_id,
-                "run.timed_out",
-                json!({
-                    "status": "timed_out",
-                    "terminal": true,
-                    "terminal_state": "timed_out",
-                    "mode": RPC_STUB_MODE,
-                    "run_id": run_id,
-                    "reason": reason,
-                }),
-            ))
-        }
-        RpcFrameKind::RunStatus => {
-            let run_id =
-                require_non_empty_payload_string(&frame.payload, "run_id", frame.kind.as_str())?;
-            Ok(build_response_frame(
-                &frame.request_id,
-                "run.status",
-                json!({
-                    "status": "inactive",
-                    "mode": RPC_STUB_MODE,
-                    "run_id": run_id,
-                    "active": false,
-                    "known": false,
-                    "terminal": false,
-                }),
-            ))
-        }
-    }
-}
-
 #[cfg(test)]
-pub fn dispatch_rpc_frame_file(path: &Path) -> Result<RpcResponseFrame> {
+pub fn dispatch_rpc_frame_file(path: &std::path::Path) -> Result<RpcResponseFrame> {
     let frame = validate_rpc_frame_file(path)?;
     dispatch_rpc_frame(&frame)
-}
-
-pub fn dispatch_rpc_raw_with_error_envelope(raw: &str) -> RpcResponseFrame {
-    match parse_rpc_frame(raw) {
-        Ok(frame) => match dispatch_rpc_frame(&frame) {
-            Ok(response) => response,
-            Err(error) => build_error_response_frame(
-                &frame.request_id,
-                classify_rpc_error_message(&error.to_string()),
-                &error.to_string(),
-            ),
-        },
-        Err(error) => {
-            let request_id =
-                best_effort_request_id_from_raw(raw).unwrap_or_else(|| "unknown".to_string());
-            build_error_response_frame(
-                &request_id,
-                classify_rpc_error_message(&error.to_string()),
-                &error.to_string(),
-            )
-        }
-    }
-}
-
-fn dispatch_rpc_raw_with_error_envelope_for_serve(
-    raw: &str,
-    state: &mut RpcServeSessionState,
-) -> Vec<RpcResponseFrame> {
-    match parse_rpc_frame(raw) {
-        Ok(frame) => match dispatch_rpc_frame_for_serve(&frame, state) {
-            Ok(responses) => responses,
-            Err(error) => vec![build_error_response_frame(
-                &frame.request_id,
-                classify_rpc_error_message(&error.to_string()),
-                &error.to_string(),
-            )],
-        },
-        Err(error) => {
-            let request_id =
-                best_effort_request_id_from_raw(raw).unwrap_or_else(|| "unknown".to_string());
-            vec![build_error_response_frame(
-                &request_id,
-                classify_rpc_error_message(&error.to_string()),
-                &error.to_string(),
-            )]
-        }
-    }
-}
-
-fn current_unix_timestamp_ms() -> u64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|duration| duration.as_millis() as u64)
-        .unwrap_or(0)
-}
-
-fn build_lifecycle_control_audit_record(
-    request_id: &str,
-    run_id: &str,
-    action: &str,
-    from_state: &str,
-    to_state: &str,
-) -> Value {
-    json!({
-        "record_type": LIFECYCLE_CONTROL_AUDIT_RECORD_TYPE_V1,
-        "schema_version": LIFECYCLE_CONTROL_AUDIT_SCHEMA_VERSION,
-        "timestamp_unix_ms": current_unix_timestamp_ms(),
-        "request_id": request_id,
-        "run_id": run_id,
-        "action": action,
-        "from_state": from_state,
-        "to_state": to_state,
-        "status": LIFECYCLE_CONTROL_STATUS_ACCEPTED,
-    })
-}
-
-fn lifecycle_control_audit_record_for_dispatch(
-    frame: &RpcFrame,
-    responses: &[RpcResponseFrame],
-) -> Option<Value> {
-    match frame.kind {
-        RpcFrameKind::RunStart => {
-            let run_id = responses
-                .first()
-                .and_then(|response| response.payload.get("run_id"))
-                .and_then(Value::as_str)?;
-            Some(build_lifecycle_control_audit_record(
-                &frame.request_id,
-                run_id,
-                LIFECYCLE_CONTROL_ACTION_RESUME,
-                "inactive",
-                "running",
-            ))
-        }
-        RpcFrameKind::RunCancel => {
-            let run_id = frame.payload.get("run_id").and_then(Value::as_str)?;
-            Some(build_lifecycle_control_audit_record(
-                &frame.request_id,
-                run_id,
-                LIFECYCLE_CONTROL_ACTION_CANCEL,
-                "running",
-                "cancelled",
-            ))
-        }
-        _ => None,
-    }
-}
-
-fn dispatch_rpc_frame_for_serve(
-    frame: &RpcFrame,
-    state: &mut RpcServeSessionState,
-) -> Result<Vec<RpcResponseFrame>> {
-    match frame.kind {
-        RpcFrameKind::CapabilitiesRequest => Ok(vec![dispatch_rpc_frame(frame)?]),
-        RpcFrameKind::RunStart => {
-            let response = dispatch_rpc_frame(frame)?;
-            let run_id = response
-                .payload
-                .get("run_id")
-                .and_then(Value::as_str)
-                .ok_or_else(|| anyhow!("rpc frame kind 'run.start' response is missing run_id"))?
-                .to_string();
-            if !state.active_run_ids.insert(run_id.to_string()) {
-                bail!(
-                    "rpc frame kind 'run.start' references duplicate active run_id '{}'",
-                    run_id
-                );
-            }
-            forget_closed_run_state(state, &run_id);
-            let prompt_chars = response
-                .payload
-                .get("prompt_chars")
-                .and_then(Value::as_u64)
-                .ok_or_else(|| {
-                    anyhow!("rpc frame kind 'run.start' response is missing prompt_chars")
-                })?;
-            let mut responses = vec![response];
-            responses.extend(build_run_start_stream_frames(
-                &frame.request_id,
-                &run_id,
-                prompt_chars,
-            ));
-            Ok(responses)
-        }
-        RpcFrameKind::RunCancel => {
-            let run_id =
-                require_non_empty_payload_string(&frame.payload, "run_id", frame.kind.as_str())?;
-            if !state.active_run_ids.remove(&run_id) {
-                bail!(
-                    "rpc frame kind 'run.cancel' references unknown run_id '{}'",
-                    run_id
-                );
-            }
-            remember_closed_run_state(state, &run_id, RpcTerminalRunState::cancelled());
-            let mut responses = vec![dispatch_rpc_frame(frame)?];
-            responses.push(build_run_cancel_stream_frame(&frame.request_id, &run_id));
-            responses.push(build_run_cancel_assistant_stream_frame(
-                &frame.request_id,
-                &run_id,
-            ));
-            Ok(responses)
-        }
-        RpcFrameKind::RunComplete => {
-            let run_id =
-                require_non_empty_payload_string(&frame.payload, "run_id", frame.kind.as_str())?;
-            if !state.active_run_ids.remove(&run_id) {
-                bail!(
-                    "rpc frame kind 'run.complete' references unknown run_id '{}'",
-                    run_id
-                );
-            }
-            remember_closed_run_state(state, &run_id, RpcTerminalRunState::completed());
-            let mut responses = vec![dispatch_rpc_frame(frame)?];
-            responses.push(build_run_complete_stream_frame(&frame.request_id, &run_id));
-            responses.push(build_run_complete_assistant_stream_frame(
-                &frame.request_id,
-                &run_id,
-            ));
-            Ok(responses)
-        }
-        RpcFrameKind::RunFail => {
-            let run_id =
-                require_non_empty_payload_string(&frame.payload, "run_id", frame.kind.as_str())?;
-            if !state.active_run_ids.remove(&run_id) {
-                bail!(
-                    "rpc frame kind 'run.fail' references unknown run_id '{}'",
-                    run_id
-                );
-            }
-            let reason = resolve_run_fail_reason(&frame.payload)?;
-            remember_closed_run_state(state, &run_id, RpcTerminalRunState::failed(&reason));
-            let mut responses = vec![dispatch_rpc_frame(frame)?];
-            responses.push(build_run_failed_stream_frame(
-                &frame.request_id,
-                &run_id,
-                &reason,
-            ));
-            responses.push(build_run_failed_assistant_stream_frame(
-                &frame.request_id,
-                &run_id,
-                &reason,
-            ));
-            Ok(responses)
-        }
-        RpcFrameKind::RunTimeout => {
-            let run_id =
-                require_non_empty_payload_string(&frame.payload, "run_id", frame.kind.as_str())?;
-            if !state.active_run_ids.remove(&run_id) {
-                bail!(
-                    "rpc frame kind 'run.timeout' references unknown run_id '{}'",
-                    run_id
-                );
-            }
-            let reason = resolve_run_timeout_reason(&frame.payload)?;
-            remember_closed_run_state(state, &run_id, RpcTerminalRunState::timed_out(&reason));
-            let mut responses = vec![dispatch_rpc_frame(frame)?];
-            responses.push(build_run_timeout_stream_frame(
-                &frame.request_id,
-                &run_id,
-                &reason,
-            ));
-            responses.push(build_run_timeout_assistant_stream_frame(
-                &frame.request_id,
-                &run_id,
-                &reason,
-            ));
-            Ok(responses)
-        }
-        RpcFrameKind::RunStatus => {
-            let run_id =
-                require_non_empty_payload_string(&frame.payload, "run_id", frame.kind.as_str())?;
-            if state.active_run_ids.contains(&run_id) {
-                return Ok(vec![build_response_frame(
-                    &frame.request_id,
-                    "run.status",
-                    json!({
-                        "status": "active",
-                        "mode": RPC_STUB_MODE,
-                        "run_id": run_id,
-                        "active": true,
-                        "known": true,
-                        "terminal": false,
-                    }),
-                )]);
-            }
-
-            if let Some(closed) = state.closed_run_states.get(&run_id) {
-                let mut payload = serde_json::Map::new();
-                payload.insert("status".to_string(), json!(closed.terminal_state));
-                payload.insert("mode".to_string(), json!(RPC_STUB_MODE));
-                payload.insert("run_id".to_string(), json!(run_id));
-                payload.insert("active".to_string(), json!(false));
-                payload.insert("known".to_string(), json!(true));
-                payload.insert("terminal".to_string(), json!(true));
-                payload.insert("terminal_state".to_string(), json!(closed.terminal_state));
-                if let Some(reason) = &closed.reason {
-                    payload.insert("reason".to_string(), json!(reason));
-                }
-                return Ok(vec![build_response_frame(
-                    &frame.request_id,
-                    "run.status",
-                    Value::Object(payload),
-                )]);
-            }
-
-            Ok(vec![build_response_frame(
-                &frame.request_id,
-                "run.status",
-                json!({
-                    "status": "inactive",
-                    "mode": RPC_STUB_MODE,
-                    "run_id": run_id,
-                    "active": false,
-                    "known": false,
-                    "terminal": false,
-                }),
-            )])
-        }
-    }
-}
-
-pub fn dispatch_rpc_ndjson_input(raw: &str) -> RpcNdjsonDispatchReport {
-    let mut responses = Vec::new();
-    let mut processed_lines = 0_usize;
-    let mut error_count = 0_usize;
-
-    for line in raw.lines() {
-        let trimmed = line.trim();
-        if trimmed.is_empty() || trimmed.starts_with('#') {
-            continue;
-        }
-        processed_lines = processed_lines.saturating_add(1);
-        let response = dispatch_rpc_raw_with_error_envelope(trimmed);
-        if response.kind == RPC_ERROR_KIND {
-            error_count = error_count.saturating_add(1);
-        }
-        responses.push(response);
-    }
-
-    RpcNdjsonDispatchReport {
-        responses,
-        processed_lines,
-        error_count,
-    }
-}
-
-pub fn serve_rpc_ndjson_reader<R, W>(mut reader: R, writer: &mut W) -> Result<RpcNdjsonServeReport>
-where
-    R: BufRead,
-    W: Write,
-{
-    let mut line = String::new();
-    let mut processed_lines = 0_usize;
-    let mut error_count = 0_usize;
-    let mut state = RpcServeSessionState::default();
-
-    loop {
-        line.clear();
-        let bytes_read = reader
-            .read_line(&mut line)
-            .context("failed to read rpc ndjson input line")?;
-        if bytes_read == 0 {
-            break;
-        }
-        let trimmed = line.trim();
-        if trimmed.is_empty() || trimmed.starts_with('#') {
-            continue;
-        }
-        processed_lines = processed_lines.saturating_add(1);
-        let responses = dispatch_rpc_raw_with_error_envelope_for_serve(trimmed, &mut state);
-        for response in responses {
-            if response.kind == RPC_ERROR_KIND {
-                error_count = error_count.saturating_add(1);
-            }
-            serde_json::to_writer(&mut *writer, &response)
-                .context("failed to serialize rpc response frame")?;
-            writer
-                .write_all(b"\n")
-                .context("failed to write rpc response delimiter")?;
-            writer
-                .flush()
-                .context("failed to flush rpc response line")?;
-        }
-    }
-
-    Ok(RpcNdjsonServeReport {
-        processed_lines,
-        error_count,
-    })
-}
-
-pub fn serve_rpc_ndjson_reader_with_lifecycle_audit<R, W, A>(
-    mut reader: R,
-    writer: &mut W,
-    lifecycle_audit_writer: &mut A,
-) -> Result<RpcNdjsonServeReport>
-where
-    R: BufRead,
-    W: Write,
-    A: Write,
-{
-    let mut line = String::new();
-    let mut processed_lines = 0_usize;
-    let mut error_count = 0_usize;
-    let mut state = RpcServeSessionState::default();
-
-    loop {
-        line.clear();
-        let bytes_read = reader
-            .read_line(&mut line)
-            .context("failed to read rpc ndjson input line")?;
-        if bytes_read == 0 {
-            break;
-        }
-        let trimmed = line.trim();
-        if trimmed.is_empty() || trimmed.starts_with('#') {
-            continue;
-        }
-        processed_lines = processed_lines.saturating_add(1);
-
-        let responses = match parse_rpc_frame(trimmed) {
-            Ok(frame) => match dispatch_rpc_frame_for_serve(&frame, &mut state) {
-                Ok(responses) => {
-                    if let Some(audit_record) =
-                        lifecycle_control_audit_record_for_dispatch(&frame, &responses)
-                    {
-                        serde_json::to_writer(&mut *lifecycle_audit_writer, &audit_record)
-                            .context("failed to serialize lifecycle audit record")?;
-                        lifecycle_audit_writer
-                            .write_all(b"\n")
-                            .context("failed to write lifecycle audit delimiter")?;
-                        lifecycle_audit_writer
-                            .flush()
-                            .context("failed to flush lifecycle audit line")?;
-                    }
-                    responses
-                }
-                Err(error) => vec![build_error_response_frame(
-                    &frame.request_id,
-                    classify_rpc_error_message(&error.to_string()),
-                    &error.to_string(),
-                )],
-            },
-            Err(error) => {
-                let request_id = best_effort_request_id_from_raw(trimmed)
-                    .unwrap_or_else(|| "unknown".to_string());
-                vec![build_error_response_frame(
-                    &request_id,
-                    classify_rpc_error_message(&error.to_string()),
-                    &error.to_string(),
-                )]
-            }
-        };
-
-        for response in responses {
-            if response.kind == RPC_ERROR_KIND {
-                error_count = error_count.saturating_add(1);
-            }
-            serde_json::to_writer(&mut *writer, &response)
-                .context("failed to serialize rpc response frame")?;
-            writer
-                .write_all(b"\n")
-                .context("failed to write rpc response delimiter")?;
-            writer
-                .flush()
-                .context("failed to flush rpc response line")?;
-        }
-    }
-
-    Ok(RpcNdjsonServeReport {
-        processed_lines,
-        error_count,
-    })
 }
 
 fn build_response_frame(request_id: &str, kind: &str, payload: Value) -> RpcResponseFrame {
